% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fit.R
\name{do_fit}
\alias{do_fit}
\title{Fit model}
\usage{
do_fit(
  data = NULL,
  x = NULL,
  y = NULL,
  col_y = NULL,
  id = NULL,
  col_id = NULL,
  overwrite = TRUE,
  export = TRUE,
  minimal = FALSE,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  ...,
  drop = TRUE,
  dir = .get_dir_data(),
  file = NULL,
  ext = "rds",
  sep = "_",
  suffix = NULL,
  seed = .get_seed(),
  .do = list(tune = NULL, fit = NULL),
  .f_import = list(tune = NULL, fit = if (minimal) {     xgboost::xgb.load } else {    
    NULL }),
  .f_export = list(tune = NULL, fit = if (minimal) {     xgboost::xgb.save } else {    
    NULL }),
  .export = list(tune = NULL, fit = NULL),
  .overwrite = list(tune = NULL, fit = NULL),
  .dir = list(tune = NULL, fit = NULL),
  .file = list(tune = "tune", fit = "fit"),
  .ext = list(tune = NULL, fit = if (minimal) {     "model" } else {     NULL }),
  .path = list(tune = NULL, fit = NULL),
  print_every_n = 10,
  fold_ids = NULL,
  n_fold = 10,
  strata = NULL,
  col_strata = NULL,
  wt = NULL,
  col_wt = NULL,
  cols_extra = NULL,
  grid_params = NULL,
  n_param = 10,
  nrounds = 2000,
  booster = "gbtree",
  objective = "reg:squarederror",
  eval_metrics = list("rmse"),
  .params = list(NULL),
  f_slice = dplyr::slice_min,
  early_stopping_rounds = 10
)
}
\description{
Tune and fit model.
}
