% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plot_shap.R
\name{do_plot_shap}
\alias{do_plot_shap}
\title{Make SHAP plots}
\usage{
do_plot_shap(
  shap,
  col_y = NULL,
  col_id = NULL,
  cols_extra = NULL,
  f_import = magick::image_read,
  f_export = save_plot,
  do = TRUE,
  overwrite = TRUE,
  export = TRUE,
  ...,
  dir = .get_dir_data(),
  file = NULL,
  ext = "png",
  sep = "_",
  prefix = "viz",
  suffix = NULL,
  .do = list(agg = NULL, swarm = NULL),
  .f_import = list(agg = NULL, swarm = NULL),
  .f_export = list(agg = purrr::partial(save_plot, width = 8, height = 8), swarm =
    purrr::partial(save_plot, width = 12, height = 8)),
  .export = list(agg = NULL, swarm = NULL),
  .overwrite = list(agg = NULL, swarm = NULL),
  .dir = list(agg = NULL, swarm = NULL),
  .file = list(agg = "agg", swarm = "swarm"),
  .ext = list(agg = NULL, swarm = NULL),
  .path = list(agg = NULL, swarm = NULL)
)
}
\description{
Make SHAP plots
}
