% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit model}
\usage{
fit(
  data = NULL,
  x = NULL,
  y = NULL,
  col_y = NULL,
  id = NULL,
  col_id = NULL,
  overwrite = TRUE,
  export = TRUE,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  ...,
  drop = TRUE,
  dir = .get_dir_data(),
  file = "fit",
  ext = "rds",
  path = NULL,
  sep = "_",
  suffix = NULL,
  print_every_n = 10,
  cols_extra = NULL,
  n_param = 10,
  nrounds = 2000,
  booster = "gbtree",
  objective = "reg:squarederror",
  eval_metrics = list("rmse"),
  .params = list(NULL),
  early_stopping_rounds = 10
)
}
\description{
Fit model.
}
