% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{generate_grid_params}
\alias{generate_grid_params}
\title{Generate a parameter grid}
\usage{
generate_grid_params(x, n_param = 10, n_col = ncol(x))
}
\description{
Generate a parameter grid given data.
}
