% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plot_shap.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot}
\usage{
save_plot(
  object,
  path = NULL,
  dir = .get_dir_data(),
  file = deparse(substitute(object)),
  ext = "png",
  height = 8,
  width = height,
  ...
)
}
\arguments{
\item{height, width}{Height and width to use for saved image, in inches.}
}
\description{
Save params
}
