% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fit.R
\name{do_fit}
\alias{do_fit}
\title{Fit model}
\usage{
do_fit(
  data = NULL,
  x = NULL,
  y = NULL,
  col_y = NULL,
  id = NULL,
  col_id = NULL,
  overwrite = TRUE,
  export = TRUE,
  ...,
  drop = TRUE,
  dir = .get_dir_data(),
  sep = "_",
  suffix = NULL,
  seed = .get_seed(),
  .overwrite = list(tune = NULL, fit = NULL),
  .export = list(tune = NULL, fit = NULL),
  .path = list(tune = NULL, fit = NULL),
  print_every_n = 10,
  fold_ids = NULL,
  n_fold = 10,
  strata = NULL,
  col_strata = NULL,
  wt = NULL,
  col_wt = NULL,
  grid_params = NULL,
  n_param = 10,
  nrounds = 2000,
  booster = "gbtree",
  objective = "reg:squarederror",
  eval_metrics = list("rmse"),
  early_stopping_rounds = 10
)
}
\description{
Tune and fit model.
}
