% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_predict.R
\name{do_predict}
\alias{do_predict}
\title{Make predictions}
\usage{
do_predict(
  data = NULL,
  fit = NULL,
  path_fit = NULL,
  x = NULL,
  y = NULL,
  col_y = NULL,
  id = NULL,
  col_id = NULL,
  overwrite = TRUE,
  export = TRUE,
  ...,
  dir = .get_dir_data(),
  sep = "_",
  suffix = NULL,
  .overwrite = list(preds = NULL, shap = NULL),
  .export = list(preds = NULL, shap = NULL),
  .path = list(preds = NULL, shap = NULL),
  cols_extra = NULL,
  f_trans = NULL
)
}
\description{
Make predictions for new data.
}
